Create Magic Square
We define a magic square to be an n x n matrix of
distinct positive integers from 1 to 2 where the sum
of any row, column, or diagonal of length n is always
equal to the same number: the magic constant. You will be given a 3 x 3 matrix s of integers in the inclusive range [1,9]. We can convert any digit a to any other digit b in the range [1,9] at cost of la- bl. Given s, convert it into a magic square at minimal cost. Print this cost on a new line.
Note: The resulting magic square must contain distinct integers in the inclusive range [1,9].
Example
$s = [[5, 3, 4], [1, 5, 8], [6, 4, 2]]
The matrix looks like this:
534
158
64 2
We can convert it to the following magic square:
8 3 4
159
672
This took three replacements at a cost of 15 8+18 9+4-77. Function Description
Complete the formingMagicSquare function in the
editor below.
formingMagicSquare has the following parameter(s):
⚫ int s[3][3]: a 3 x 3 array of integers
Returns
⚫int: the minimal total cost of converting the input square to a magic square
Input Format
Each of the 3 lines contains three space-separated integers of row s[i].
Constraints
• s[i][j] € [1,9]
Sample Input 0
492
357
8 15
Sample Output 0 1
Explanation 0
If we change the bottom right value, s[2][2], from 5 to 6 at a cost of 165 = 1, s becomes a magic square at the minimum possible cost.
Sample Input 1
48 2
457
61 6
Sample Output 1
Explanation 1
Using O-based indexing, if we make
• s[0][1]->9 at a cost of |9-8] = 1
s[1][0]->3 at a cost of 3 - 4 = 1
. s[2][0]->8 at a cost of 18 - 6 = 2,
then the total cost will be 1+1+2=4.